⚡️ [STYLE] 홈 input, map modal 스타일 통일
recommend -> home 이동
